Move dashboard and helpdesk test into their own files
