+++ app/tests/dashboard.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from "@playwright/test";
 // ---
-import { url, frf2024MongoId } from "./config";
+import { url } from "./config";
 
 test.describe("Dashboard", () => {
   test.beforeEach(async ({ page }) => {
@@ -28,26 +28,6 @@ test.describe("Dashboard", () => {
     await page.getByLabel("Rebate Year:").selectOption("2024");
 
     const table = page.getByLabel("Your 2024 Rebate Forms");
-    await expect(table).toBeVisible();
-  });
-});
-
-test.describe("Helpdesk", () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto(url);
-
-    const heading = page.getByRole("heading", { name: "Your Rebate Forms" });
-    await expect(heading).toBeVisible();
-  });
-
-  test("Submission Results", async ({ page }) => {
-    const root = page.locator("#root");
-
-    await root.getByRole("link", { name: "Helpdesk" }).click();
-    await root.getByLabel("Search submissions by ID").fill(frf2024MongoId);
-    await root.getByRole("button", { name: "Search" }).click();
-
-    const table = root.getByLabel("Submission Search Results");
     await expect(table).toBeVisible();
   });
 });
